requirements.txt
ultralytics==8.0.196
opencv-python==4.8.1.78
matplotlib==3.7.2
pillow==10.0.0
numpy==1.24.3
